# openapi/openapi0008.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "DripRate Server API",
    "description": "API for DripRate Server application",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:13000/"
    }
  ],
  "paths": {
...
    "/api/Wardrobe": {
      "post": {
        "tags": [
          "Wardrobe"
        ],
        "operationId": "Wardrobe_Add",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Name": {
                    "type": "string"
                  },
                  "Brand": {
                    "type": "string"
                  },
                  "StoreLink": {
                    "type": "string"
                  },
                  "EstimatedPrice": {
                    "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
                    "type": "number",
                    "format": "double"
                  },
                  "Photo": {
                    "$ref": "#/components/schemas/IFormFile"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Wardrobe_Add_500ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Wardrobe_Add_500ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Wardrobe_Add_400ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Wardrobe_Add_400ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Wardrobe_Add_401ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Wardrobe_Add_401ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "get": {
        "tags": [
          "Wardrobe"
        ],
        "operationId": "Wardrobe_GetWardrobe",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "newest"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClothResponseDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClothResponseDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClothResponseDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Wardrobe_GetWardrobe_400ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Wardrobe_GetWardrobe_400ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Wardrobe_GetWardrobe_401ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Wardrobe_GetWardrobe_401ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Wardrobe/{id}": {
      "delete": {
        "tags": [
          "Wardrobe"
        ],
        "operationId": "Wardrobe_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Wardrobe_Delete_404ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Wardrobe_Delete_404ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Wardrobe_Delete_403ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Wardrobe_Delete_403ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Wardrobe_Delete_400ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Wardrobe_Delete_400ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Wardrobe_Delete_401ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Wardrobe_Delete_401ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "ClothResponseDto": {
        "required": [
          "id",
          "name",
          "brand",
          "photoUrl",
          "storeLink",
          "estimatedPrice"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "brand": {
            "type": [
              "null",
              "string"
            ]
          },
          "photoUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "storeLink": {
            "type": [
              "null",
              "string"
            ]
          },
          "estimatedPrice": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          }
        }
      },
...
      "IFormFile": {
        "type": "string",
        "format": "binary"
      },
...
      "Wardrobe_Add_400ErrorCode": {
        "enum": [
          "General.Validation"
        ],
        "type": "string"
      },
      "Wardrobe_Add_401ErrorCode": {
        "enum": [
          "Auth.Unauthorized"
        ],
        "type": "string"
      },
      "Wardrobe_Add_500ErrorCode": {
        "enum": [
          "File.ProcessingFailed"
        ],
        "type": "string"
      },
      "Wardrobe_Delete_400ErrorCode": {
        "enum": [
          "General.Validation"
        ],
        "type": "string"
      },
      "Wardrobe_Delete_401ErrorCode": {
        "enum": [
          "Auth.Unauthorized"
        ],
        "type": "string"
      },
      "Wardrobe_Delete_403ErrorCode": {
        "enum": [
          "Cloth.Forbidden"
        ],
        "type": "string"
      },
      "Wardrobe_Delete_404ErrorCode": {
        "enum": [
          "Cloth.NotFound"
        ],
        "type": "string"
      },
      "Wardrobe_GetWardrobe_400ErrorCode": {
        "enum": [
          "General.Validation"
        ],
        "type": "string"
      },
      "Wardrobe_GetWardrobe_401ErrorCode": {
        "enum": [
          "Auth.Unauthorized"
        ],
        "type": "string"
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}